requirement.txt
torch
numpy
tqdm
scikit-learn
colorama
transformers
faiss-gpu
emoji
sentence-transformers